Trying to fix type issue

+++ src/score.py
@@ -30,7 +30,7 @@ def forward_pass(net, in_, target, generator=None, net_weights=None, gen_weights
         noise = Variable(torch.FloatTensor(np.random.normal(0, 1, (in_.shape[0], generator.latent_dim)))).cuda(async=True)
         one_hot_targets = torch.nn.functional.one_hot(target_var, generator.num_classes)
         fake_input_var = generator.forward(one_hot_targets, noise, gen_weights)
-        fake_target_var = Variable(torch.ones(target.shape, dtype=torch.float32) * generator.num_classes).cuda(async=True)
+        fake_target_var = Variable(torch.ones(target.shape, dtype=torch.int64) * generator.num_classes).cuda(async=True)
 
         print("Hey inside generator loop and fake target var is")
         print(fake_target_var)
